requirements.txt
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
PyQt5-stubs==5.15.2.0
pyqt5-tools==5.15.4.3.2
